Move commands to client/commands
