+++ client/commands/common.go
