+++ client/commands/init.go
