+++ client/commands/proofs/root.go
