+++ client/commands/proofs/tx.go
@@ -3,7 +3,7 @@ package proofs
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/tendermint/basecoin/commands"
+	"github.com/tendermint/basecoin/client/commands"
 	"github.com/tendermint/light-client/proofs"
 )
