+++ client/commands/rpc/helpers.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/tendermint/basecoin/commands"
+	"github.com/tendermint/basecoin/client/commands"
 
 	"github.com/tendermint/tendermint/rpc/client"
 )
