+++ client/commands/rpc/secure.go
@@ -3,7 +3,7 @@ package rpc
 import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/tendermint/basecoin/commands"
+	"github.com/tendermint/basecoin/client/commands"
 )
 
 func init() {
