+++ cmd/basecli/main.go
@@ -8,12 +8,12 @@ import (
 
 	"github.com/tendermint/abci/version"
 	keycmd "github.com/tendermint/go-crypto/cmd"
-	"github.com/tendermint/basecoin/commands"
-	"github.com/tendermint/basecoin/commands/proofs"
-	"github.com/tendermint/basecoin/commands/proxy"
-	rpccmd "github.com/tendermint/basecoin/commands/rpc"
-	"github.com/tendermint/basecoin/commands/seeds"
-	"github.com/tendermint/basecoin/commands/txs"
+	"github.com/tendermint/basecoin/client/commands"
+	"github.com/tendermint/basecoin/client/commands/proofs"
+	"github.com/tendermint/basecoin/client/commands/proxy"
+	rpccmd "github.com/tendermint/basecoin/client/commands/rpc"
+	"github.com/tendermint/basecoin/client/commands/seeds"
+	"github.com/tendermint/basecoin/client/commands/txs"
 	"github.com/tendermint/tmlibs/cli"
 
 	bcmd "github.com/tendermint/basecoin/cmd/basecli/commands"
