+++ docs/guide/counter/cmd/countercli/main.go
@@ -6,11 +6,11 @@ import (
 	"github.com/spf13/cobra"
 
 	keycmd "github.com/tendermint/go-crypto/cmd"
-	"github.com/tendermint/basecoin/commands"
-	"github.com/tendermint/basecoin/commands/proofs"
-	"github.com/tendermint/basecoin/commands/proxy"
-	"github.com/tendermint/basecoin/commands/seeds"
-	"github.com/tendermint/basecoin/commands/txs"
+	"github.com/tendermint/basecoin/client/commands"
+	"github.com/tendermint/basecoin/client/commands/proofs"
+	"github.com/tendermint/basecoin/client/commands/proxy"
+	"github.com/tendermint/basecoin/client/commands/seeds"
+	"github.com/tendermint/basecoin/client/commands/txs"
 	"github.com/tendermint/tmlibs/cli"
 
 	bcmd "github.com/tendermint/basecoin/cmd/basecli/commands"
