+++ modules/coin/commands/tx.go
@@ -4,8 +4,8 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/tendermint/basecoin/commands"
-	txcmd "github.com/tendermint/basecoin/commands/txs"
+	"github.com/tendermint/basecoin/client/commands"
+	txcmd "github.com/tendermint/basecoin/client/commands/txs"
 
 	"github.com/tendermint/basecoin"
 	bcmd "github.com/tendermint/basecoin/cmd/basecli/commands"
